+++ js/data.js
@@ -1,0 +1,63 @@
+export const minGeneralNumber = 1;
+export const maxGeneralNumber = 25;
+export const minLikesNumber = 15;
+export const maxLikesNumber = 200;
+export const minCommentNumber = 0;
+export const maxCommentNumber = 30;
+export const minAvatarPhoto = 1;
+export const maxAvatarPhoto = 6;
+
+export const descriptions = [
+  'Прекрасный момент',
+  'Просто день',
+  'На память',
+  'Легкий кадр',
+  'Обычный день',
+  'Просто фото',
+  'Мгновение жизни',
+  'Время остановилось',
+  'Тихий уголок',
+  'Без слов',
+  'Просто так',
+  'Немного света',
+  'Мир вокруг',
+  'Взгляд в объектив',
+  'Пауза на секунду',
+  'Теплый день',
+  'Просто настроение',
+  'Спокойствие и тишина',
+  'Время для себя',
+  'Маленькое счастье',
+  'Свет и тень',
+  'Вдохновение рядом',
+  'Обычный кадр',
+  'Просто здесь и сейчас',
+  'Момент тишины',
+  'Легкость бытия',
+  'Взгляд в небо',
+  'Тишина вокруг',
+  'Небольшое приключение',
+  'Просто жизнь'
+];
+
+export const messages = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export const names = [
+  'Алексей',
+  'Марина',
+  'Игорь',
+  'Екатерина',
+  'Дмитрий',
+  'Ольга',
+  'Сергей',
+  'Анна',
+  'Владимир',
+  'Татьяна'
+];
